Moved and refactored a bunch of stuff
