--- src/index.ts
@@ -1,5 +1,0 @@
-export { kint } from './kint';
-export { k } from './schemaHelpers';
-
-import { kint } from './kint';
-export default kint;
