+++ src/middleware/models/PostProcessingMiddlewareTuple.ts
@@ -1,0 +1,6 @@
+import { PostProcessingMiddleware } from './PostProcessingMiddleware';
+
+export type PostProcessingMiddlewareTuple = PostProcessingMiddleware<
+	any,
+	any
+>[];
