+++ src/middleware/models/PreprocessingMiddleware.ts
@@ -1,0 +1,9 @@
+import { RawKintRequest } from '../../models/RawKintRequest';
+
+export type PreprocessingMiddleware<
+	Config,
+	OutputRequest,
+	InputRequest = RawKintRequest
+> = {
+	preProcess: (request: InputRequest, config: Config) => OutputRequest;
+};
