+++ src/middleware/models/StrictPreprocessingMiddleware.ts
@@ -1,0 +1,8 @@
+import { RawKintRequest } from '../../models/RawKintRequest';
+import { PreprocessingMiddleware } from './PreprocessingMiddleware';
+
+export type StrictPreprocessingMiddleware<
+	Config,
+	OutputRequest extends RawKintRequest,
+	InputRequest = RawKintRequest
+> = PreprocessingMiddleware<Config, OutputRequest, InputRequest>;
