+++ src/middleware/utils/PostProcessorCatchTypes.ts
@@ -1,0 +1,12 @@
+import { PostProcessingMiddlewareTuple } from '../models/PostProcessingMiddlewareTuple';
+import { PostProcessingMiddleware } from '../models/PostProcessingMiddleware';
+
+export type PostProcessorCatchTypes<
+	ConcretePostProcessingMiddlewareTuple extends PostProcessingMiddlewareTuple
+> =
+	ConcretePostProcessingMiddlewareTuple[number] extends PostProcessingMiddleware<
+		any,
+		infer CatchType
+	>
+		? CatchType
+		: never;
