--- src/models/GenericKintRequest.ts
@@ -1,5 +1,0 @@
-export interface GenericKintRequest<RequestBody, UrlParams, QueryParams> {
-	requestBody: RequestBody;
-	urlParams: UrlParams;
-	queryParams: QueryParams;
-}
