+++ src/models/RawKintResponse.ts
@@ -1,0 +1,1 @@
+export type RawKintResponse = {};
