+++ src/utils/AppendTuple.ts
@@ -1,0 +1,1 @@
+export type AppendTuple<T extends any[], E> = [...T, E];
