+++ src/utils/mergeConfigs.ts
@@ -1,0 +1,7 @@
+export function mergeConfigs<Config>(
+	defaultConfig: Config,
+	userConfig: Partial<Config>
+): Config {
+	// TODO: MERGE CONFIGS
+	return defaultConfig;
+}
